chore: refactors state flows in auth repository
