+++ app/src/main/java/com/codemote/puppytimev2/core/AuthState.kt
@@ -1,4 +1,4 @@
-package com.codemote.puppytimev2.common
+package com.codemote.puppytimev2.core
 
 enum class AuthState {
     SIGNED_IN, SIGNED_OUT, LOADING
