+++ app/src/main/java/com/codemote/puppytimev2/core/RemoteResult.kt
@@ -1,4 +1,4 @@
-package com.codemote.puppytimev2.common
+package com.codemote.puppytimev2.core
 
 sealed interface RemoteResult<T>
 data class Success<T>(val data: T) : RemoteResult<T>
