+++ app/src/main/java/com/codemote/puppytimev2/core/constants/ScreenConstants.kt
@@ -1,4 +1,4 @@
-package com.codemote.puppytimev2.common.constants
+package com.codemote.puppytimev2.core.constants
 
 const val ACCOUNT_SCREEN = "account"
 const val HOME_SCREEN = "home"
