+++ app/src/main/java/com/codemote/puppytimev2/domain/repository/AuthRepository.kt
@@ -1,15 +1,14 @@
 package com.codemote.puppytimev2.domain.repository
 
-import com.codemote.puppytimev2.common.AuthRemoteResult
-import com.codemote.puppytimev2.common.AuthState
+import com.codemote.puppytimev2.core.AuthRemoteResult
+import com.codemote.puppytimev2.core.AuthState
 import com.google.firebase.auth.FirebaseUser
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.Flow
 
 interface AuthRepository {
     suspend fun createUser(email: String, name: String, password: String): AuthRemoteResult
-    fun getCurrentUser(viewModelScope: CoroutineScope): StateFlow<FirebaseUser?>
-    fun getSignedState(viewModelScope: CoroutineScope): StateFlow<AuthState>
+    fun getCurrentUser(): Flow<FirebaseUser?>
+    fun getSignedState(): Flow<AuthState>
     suspend fun authenticateUser(email: String, password: String): AuthRemoteResult
     fun signOut(): AuthRemoteResult
 }
